Relocate useSessions.ts and useSessions.test.ts into hooks folder
