+++ src/entities/sessions/hooks/useSessions/useSessions.test.ts
@@ -1,11 +1,11 @@
 import { renderHook } from "@testing-library/react";
-import { sessionsMock } from "../mocks/sessionsMock";
+import { sessionsMock } from "../../mocks/sessionsMock";
 
 import useSessions from "./useSessions";
-import { wrapWithProviders } from "../../../utils/testUtils";
-import { SessionsStructure } from "../types";
-import { server } from "../../../mocks/node";
-import { errorHandlers } from "../../../mocks/handlers";
+import { wrapWithProviders } from "../../../../utils/testUtils";
+import { SessionsStructure } from "../../types";
+import { server } from "../../../../mocks/node";
+import { errorHandlers } from "../../../../mocks/handlers";
 
 describe("Given a useSessions function", () => {
   describe("When it calls the getSessions function", () => {
